requirements.txt
colormath==3.0.0
opencv_contrib_python==4.5.4.60
pandas==1.3.5
Pillow==9.1.1
scikit_learn==1.1.1
